perf: optimize phase-distortion algos with custom pow01 approximation
- Implement piecewise pow01() for pinch, sync, twist, bend algorithms
- Specialized approximation for [0,1] phase range avoids f32↔f64 conversion overhead
- Tested fastmaths crate; caused regressions due to type conversion costs
- Revert to libm::sinf for base waveforms (fastmaths::sin also had overhead)

Performance improvements:
- Pinch: 617.4 → 429.5 ns/sample (30% gain)
- Sync: 622.4 → 443.3 ns/sample (29% gain)
- Twist: 719.7 → 455.0 ns/sample (37% gain)
- Bend: ~640 → 556.1 ns/sample (13% gain)

Remove fastmaths dependency; custom approximation is faster for this use case.

+++ packages/cosmo-synth-engine/src/generators/mod.rs
@@ -220,7 +220,7 @@ fn sample_base_wave(base_waveform: BaseWaveform, phase: f32) -> f32 {
     };
     match base_waveform {
         BaseWaveform::Cosine => -libm::cosf(TWO_PI * p),
-        BaseWaveform::Sine => libm::sinf(TWO_PI * p),
+        BaseWaveform::Sine => libm::sinf(TWO_PI as f32 * p),
         BaseWaveform::Triangle => 1.0 - 4.0 * libm::fabsf(p - 0.5),
         BaseWaveform::Saw => p * 2.0 - 1.0,
         BaseWaveform::Square => {
